* Folder structure reorganized. * Documentation updated. * Introduced configuration. * Upgraded dependencies * Made some minor improvements
