--- src/Domain/Menu/Menu.php
@@ -1,20 +1,0 @@
-<?php
-/**
- * The file is part of the "webifycms/ext-admin", WebifyCMS extension package.
- *
- * @see https://webifycms.com/extension/admin
- *
- * @copyright Copyright (c) 2023 WebifyCMS
- * @license https://webifycms.com/extension/admin/license
- * @author Mohammed Shifreen <mshifreen@gmail.com>
- */
-declare(strict_types=1);
-
-namespace Webify\Admin\Domain\Menu;
-
-/**
- * Class Menu.
- */
-final class Menu
-{
-}
